pyproject.toml
[project]
name = "rag-learning-lab"
version = "0.1.0"
description = "An open-source project for learning RAG in public"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    # Core orchestration
    "llama-index>=0.14",
    "llama-index-vector-stores-chroma>=0.5",
    "llama-index-llms-ollama>=0.9",
    "llama-index-embeddings-huggingface>=0.6",
    # Vector stores
    "chromadb>=0.5",
    "lancedb>=0.14",
    # Embeddings & reranking
    "sentence-transformers>=3.0",
    "rank-bm25>=0.2",
    # Document parsing
    "pdfplumber>=0.11",
    "python-docx>=1.0",
    "beautifulsoup4>=4.12",
    "markdown>=3.5",
    # Evaluation
    "ragas>=0.2",
    "deepeval>=1.0",
    # Experiment tracking
    "mlflow>=2.15",
    # Prompt templates
    "jinja2>=3.1",
    # Data handling & utilities
    "pyyaml>=6.0",
    "numpy>=1.26",
    "pandas>=2.1",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    # API & serving
    "fastapi>=0.115",
    "uvicorn>=0.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "pre-commit>=3.8",
    "ipykernel>=6.29",
    "jupyterlab>=4.3",
    "matplotlib>=3.9",
    "scikit-learn>=1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "eval: evaluation tests",
    "security: security tests",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[dependency-groups]
dev = [
    "fpdf2>=2.8.7",
]
